Ensure process is defined in version check
To allow for user agents which contain "electron" within their name but do not run within a node.js context, we need to check if the process object is defined at all before accessing its properties.

In particular, this is required to run diagrams.net within a Visual Studio Code webview extension.